You Now Create Kit Features As Specific Types In The Actor Sheet
You can now create features specifically as weapons, talents, abilities, and passives. The old "Kit Feature" bar has been removed as it is now redundant. All of this is done in one function, however it is proofed to work without the data-type attribute set. Use data-type attribute to create a kit piece already of a specific type.